requirements.txt
langchain==0.1.0
langchain-openai==0.0.2
python-dotenv==1.0.0
langchain-community==0.0.10
flask==3.0.0
flask-cors==4.0.0
langchain-groq
